Don't let highs affect perceived bass loudness
Oh no! This was letting the rest of the bands of the song affect how loud the algorithm thought the bass was. We don't care about them! We just want to know when kicks are happening!

This eliminates all but the bands that matter from hzHistory, and therefore from affecting the result of getFrequencyRatio.