Use timers instead of polling for new lyrics
This turns the render function from a sync function that runs every frame to an async function that uses timeouts to update the lyrics. should be a lot more efficient this way.

Also incorporates new logic that should hopefully make it scrub-proof! Before if you rewound or fast-forwarded the song it would get confused. That shouldn't happen anymore.